Update WalletBalance to Decimal and refactor data mapping
Changed WalletBalance from double? to Decimal? for accuracy in Users, UserDTO, and UserInfoDTO. Updated constructors accordingly. Introduced MapRow method in UsersData.cs for safer and consistent SqlDataReader to DTO mapping. Replaced direct DTO construction with MapRow usage.

+++ BreadApp_BL/Users.cs
@@ -17,7 +17,7 @@ namespace BreadApp_BL
         public int? FamilyNumber { get; set; }
         public string? Phone { get; set; }
         public string? PasswordHash { get; set; }
-        public double? WalletBalance { get; set; }
+        public Decimal? WalletBalance { get; set; }
         public string? WifeHusbNational { get; set; }
         public bool? IsActive { get; set; }
         public DateTime? CreatedAt { get; set; }
@@ -42,7 +42,7 @@ namespace BreadApp_BL
         }
         private Users(int UserID ,Guid PublicID , string NationalNumber , string FirstName , string SecondName , string LastName , 
             DateTime DateOfBirth , bool MaritalStatus , int FamilyNumber , string Phone , string PasswordHash
-            , double WalletBalance , string WifeHusbNational , bool IsActive , DateTime CreatedAt)
+            , Decimal WalletBalance , string WifeHusbNational , bool IsActive , DateTime CreatedAt)
         {
             this.UserID = UserID;
             this.PublicID = PublicID;
